improved getStructureLinks.m
Updated utility function for identifying pre-synaptic neuron IDs.
First, some housekeeping as this was not updated for the new JSON
decoder. Added a new output argument (synapseIDs) which is a vector
of the presynaptic synapse IDs matching the vector of posysynaptic
neuron IDs (linkedIDs).

There is still an error in the third output, a graph representation
of the network, which occurs when 2 or less nodes are in the graph.
This should be removed from the code in the future. Additionally,
bidirectional synapses are not currently supported.